Style: yAxis 폰트 사이즈 변경 , 그외 코드 변경
