+++ src/router/router.tsx
@@ -11,8 +11,8 @@ function Router() {
     <BrowserRouter>
       <Routes>
         <Route path="" element={<Layout />}>
-          <Route index element={<MediaStatus />} />
-          <Route path="dashboard" element={<MediaStatus />} />
+          <Route index element={<Dashboard />} />
+          <Route path="dashboard" element={<Dashboard />} />
           <Route path="management" element={<Management />} />
           <Route path="create" element={<CreateAd />} />
           <Route path="*" element={<div>페이지가 없습니다</div>} />
